fix: resolve clippy warnings across codebase (#309)
- Use &Path instead of &PathBuf in build.rs
- Fix doc comment formatting in audio.rs, hint.rs, render.rs
- Remove redundant wildcard pattern in event.rs
- Remove unnecessary type casts in filesystem.rs, messagebox.rs, video.rs
- Add is_empty() methods to Buffer and TransferBuffer
- Allow doc_lazy_continuation lint for generated hint names

+++ src/sdl3/filesystem.rs
@@ -303,7 +303,7 @@ bitflags! {
     #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
     pub struct GlobFlags: u32 {
         const NONE = 0;
-        const CASEINSENSITIVE = sys::filesystem::SDL_GLOB_CASEINSENSITIVE.0 as u32;
+        const CASEINSENSITIVE = sys::filesystem::SDL_GLOB_CASEINSENSITIVE.0;
     }
 }
